fix: rename team-projects-states.ts to team-project-state.ts matching the class name
